feat: 簡易版signin/out
