+++ app/(authorized)/(tabs)/account/_layout.tsx
@@ -1,10 +1,8 @@
 import { FontAwesome } from "@expo/vector-icons";
 import { Link, Stack } from "expo-router";
-import { Pressable, useColorScheme } from "react-native";
+import { Pressable } from "react-native";
 
 export default function Layout() {
-  const colorScheme = useColorScheme();
-
   return (
     <Stack>
       <Stack.Screen
