+++ app/(authorized)/(tabs)/create-card/_layout.tsx
