+++ app/(authorized)/(tabs)/home/[id]/index.tsx
