+++ app/(authorized)/(tabs)/home/typography/index.tsx
