+++ app/(authorized)/(tabs)/notification/_layout.tsx
