--- components/__tests__/StyledText-test.js
@@ -1,9 +1,0 @@
-import renderer from "react-test-renderer";
-
-import { MonoText } from "../StyledText";
-
-it("renders correctly", () => {
-  const tree = renderer.create(<MonoText>Snapshot test!</MonoText>).toJSON();
-
-  expect(tree).toMatchSnapshot();
-});
